Fix most clippy pedantic warnings.

+++ jit/src/lib.rs
@@ -8,13 +8,13 @@ pub mod runtime;
 pub mod test_roms;
 
 unsafe extern "C" {
-    fn console_log_glue(buffer: *const u8, len: u32);
+    fn console_log_glue(buffer: *const u8, len: usize);
 }
 
 /// This function is expensive to call, it should have no usages in the codebase unless it's actively being used for debugging.
 pub fn console_log(message: &str) {
     unsafe {
-        console_log_glue(message.as_ptr(), message.len() as u32);
+        console_log_glue(message.as_ptr(), message.len());
     }
 }
